+++ integration/lanesManager/init.lua
@@ -132,7 +132,7 @@ function multi.InitSystemThreadErrorHandler()
 			thread.yield()
 			_,data = __ConsoleLinda:receive(0, "Q")
 			if data then
-				print(data[1])
+				--print(data[1])
 			end
 			for i = #threads, 1, -1 do
 				temp = threads[i]